multilingual-product-api/src/products/products.controller.ts
import { Controller, Post, Body, Get, Query } from '@nestjs/common';
import { ProductsService } from './products.service';
import { CreateProductDto } from './dto/create-product.dto';

@Controller('products')
export class ProductsController {
  constructor(private readonly productsService: ProductsService) { }

  @Post()
  async create(@Body() createProductDto: CreateProductDto) {
    return this.productsService.create(createProductDto);
  }

  @Get('search')
  async search(
    @Query('query') query: string,
    @Query('language') language: string = 'en', // Default to English if no language is provided
    @Query('page') page: number = 1,
    @Query('limit') limit: number = 10,
  ) {
    return this.productsService.search(query, language, page, limit);
  }
}
